+++ sdk/transcriber_models.py
@@ -188,7 +188,7 @@ class APIWhisperSTTModel(STTModelInterface):
         if config["api_key"] is None:
             raise Exception("Attempt to create Open AI Whisper STT Model without an api key.")  # pylint: disable=W0719
         print('[INFO] Using Open AI Whisper API for transcription.')
-        self.client = openai.OpenAI(api_key=config["api_key"])
+        self.stt_client = openai.OpenAI(api_key=config["api_key"])
         # lang parameter is not required for API invocation. This exists solely
         # to support --api option from command line.
         # A better solution is to create a base class for APIWhisperSTTModel,
@@ -205,7 +205,7 @@ class APIWhisperSTTModel(STTModelInterface):
         """
         try:
             with open(wav_file_path, "rb") as audio_file:
-                result = self.client.audio.transcriptions.create(model='whisper-1', file=audio_file)
+                result = self.stt_client.audio.transcriptions.create(model='whisper-1', file=audio_file)
         except Exception as exception:
             print(exception)
             return ''
